requirements.txt
argparse
pymongo
celery
redis
Pillow